feat: add about us component
