+++ src/components/about-us/AboutUs.css
@@ -1,0 +1,19 @@
+.decagon {
+  transform: rotate(-90deg);
+  clip-path: polygon(
+    34.54915% 2.44717%,
+    65.45085% 2.44717%,
+    90.45085% 20.61074%,
+    100% 50%,
+    90.45085% 79.38926%,
+    65.45085% 97.55283%,
+    34.54915% 97.55283%,
+    9.54915% 79.38926%,
+    0% 50%,
+    9.54915% 20.61074%
+  );
+}
+
+.shadow-all {
+  box-shadow: #130a2e0d 0px 4px 20px;
+}
