+++ tailwind.config.js
@@ -2,7 +2,12 @@
 export default {
   content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      container: {
+        mx: "auto",
+        center: true,
+      },
+    },
   },
   plugins: [],
 };
